Initial release
This is a script that runs with the Create 1/2 Roomba which allows it to move as dictated by a PS5 Dualsense controller. Items can be obtained through the activation of the 
wheeldrop(s) which grant bonuses to the player's movement (e.g., increased speed). In other words, it mimics the basic capabilities of Mario Kart.